feat(table): add table scroll
