+++ core/components/organisms/grid/GridBody.tsx
@@ -1,50 +1,28 @@
 import * as React from 'react';
+// import VirtualScroll from 'react-dynamic-virtual-scroll';
 import { GridRow } from './GridRow';
 import { GridState, onSelectFn, Schema, updatePrevPageInfoFunction } from './Grid';
+// import { GridProps } from '@/index.type';
 import GridContext from './GridContext';
 import styles from '@css/components/grid.module.css';
-import { GridProps } from '@/index.type';
-import VirtualScroll from './VirtualScroll';
-import { ProgressBar } from '@/index';
 
 export interface GridBodyProps {
   schema: Schema;
   onSelect: onSelectFn;
   prevPageInfo: GridState['prevPageInfo'];
   updatePrevPageInfo: updatePrevPageInfoFunction;
-  virtualScrollOptions: GridProps['virtualScrollOptions'];
 }
 
 export const GridBody = (props: GridBodyProps) => {
   const context = React.useContext(GridContext);
 
-  const {
-    data,
-    ref,
-    loading,
-    error,
-    withPagination,
-    page,
-    pageSize,
-    totalRecords,
-    errorTemplate,
-    size,
-    // updateData,
-    updateVirtualData,
-  } = context;
+  const { data, ref, loading, error, withPagination, page, pageSize, totalRecords, errorTemplate } = context;
 
   if (!loading && error) {
     return errorTemplate ? (typeof errorTemplate === 'function' ? errorTemplate({}) : errorTemplate) : null;
   }
 
-  const { schema, prevPageInfo, updatePrevPageInfo, onSelect, virtualScrollOptions } = props;
-
-  const { preFetchRows, buffer, visibleRows, loadMoreThreshold, onScroll } = virtualScrollOptions;
-
-  const [isLoadingMore, setIsLoadingMore] = React.useState(false);
-  const [currentPage, setCurrentPage] = React.useState(1);
-  const [hasMoreData, setHasMoreData] = React.useState(true);
-  // const [scrollTop, setScrollTop] = React.useState(0);
+  const { schema, prevPageInfo, updatePrevPageInfo, onSelect } = props;
 
   React.useEffect(() => {
     const gridBodyEl = ref!.querySelector('.Grid-body');
@@ -57,8 +35,6 @@ export const GridBody = (props: GridBodyProps) => {
         gridBodyEl.scrollLeft = gridHeadEl.scrollLeft;
       });
     }
-
-    fetchNextRows();
 
     return () => {
       if (gridBodyEl) {
@@ -77,147 +53,30 @@ export const GridBody = (props: GridBodyProps) => {
     ? Math.min(totalRecords, pageSize)
     : totalRecords;
 
-  console.log(
-    'ttttt grid body',
-    data,
-    'dataLength',
-    dataLength,
-    'totalRecords',
-    totalRecords,
-    'page',
-    page,
-    'pageSize',
-    pageSize
-  );
-
-  const renderRow = React.useCallback(
-    (rowIndex: number, item?: object) => {
-      return (
-        <GridRow
-          key={rowIndex}
-          rowIndex={rowIndex}
-          data={!item ? data[rowIndex] : item}
-          schema={schema}
-          onSelect={onSelect}
-        />
-      );
-    },
-    [data, schema, onSelect]
-  );
-
-  // const renderRow = React.useCallback(
-  //   (rowIndex: number, item: object) => {
-  //     return (
-  //       <GridRow
-  //         key={rowIndex}
-  //         rowIndex={rowIndex}
-  //         data={!item ? data[rowIndex] : item}
-  //         schema={schema}
-  //         onSelect={onSelect}
-  //       />
-  //     );
-  //   },
-  //   [data, schema, onSelect]
-  // );
-
-  // const getArrayList = () => {
-  //   if (loading && !data.length) {
-  //     return [...Array(dataLength).map((x) => x)];
-  //   }
-  //   return data;
-  // };
-
-  const minRowHeight: Record<GridProps['size'], number> = {
-    comfortable: 40,
-    standard: 40,
-    compressed: 32,
-    tight: 24,
+  const renderRow = (rowIndex: number, item: object) => {
+    return (
+      <GridRow
+        key={rowIndex}
+        rowIndex={rowIndex}
+        data={!item ? data[rowIndex] : item}
+        schema={schema}
+        onSelect={onSelect}
+      />
+    );
   };
 
-  // const fetchNextRows = async () => {
-  //   console.log('<<<end reached>>>', updateVirtualData, isLoadingMore);
-
-  //   if (updateVirtualData && !isLoadingMore && hasMoreData) {
-  //     // const gridBodyEl = ref!.querySelector('.Grid-body');
-  //     // if (gridBodyEl) {
-  //     //   setScrollTop(gridBodyEl.scrollTop);
-  //     // }
-
-  //     setIsLoadingMore(true);
-  //     try {
-  //       const dataList = await updateVirtualData({ page: currentPage + 1, pageSize });
-  //       console.log('<<<dataListaa>>>', dataList);
-  //       if (dataList.length === 0) {
-  //         setHasMoreData(false);
-  //       }
-  //       setCurrentPage((prevPage) => prevPage + 1);
-  //     } finally {
-  //       setIsLoadingMore(false);
-  //     }
-  //   }
-  // };
-
-  const fetchNextRows = React.useCallback(async () => {
-    console.log('<<<end reached>>>', 'hasMoreData', hasMoreData);
-    if (updateVirtualData && !isLoadingMore && hasMoreData) {
-      setIsLoadingMore(true);
-      try {
-        const dataList = await updateVirtualData({ page: currentPage + 1, preFetchRows });
-        if (dataList.length === 0) {
-          setHasMoreData(false);
-        }
-        setCurrentPage((prevPage) => prevPage + 1);
-      } finally {
-        setIsLoadingMore(false);
-      }
+  const getArrayList = () => {
+    if (loading && !data.length) {
+      return [...Array(dataLength).map((x) => x)];
     }
-  }, [updateVirtualData, isLoadingMore, hasMoreData, currentPage, preFetchRows]);
-
-  // React.useEffect(() => {
-  //   const gridBodyEl = ref!.querySelector('.Grid-body');
-  //   if (gridBodyEl) {
-  //     gridBodyEl.scrollTop = scrollTop;
-  //   }
-  // }, [scrollTop, data]);
-
-  const memoizedVirtualScroll = React.useMemo(
-    () => (
-      <VirtualScroll
-        className={styles['Grid-body']}
-        buffer={buffer}
-        length={visibleRows}
-        minItemHeight={minRowHeight[size]}
-        totalLength={dataLength}
-        renderItem={renderRow}
-        onScroll={onScroll}
-        loadMoreThreshold={loadMoreThreshold}
-        onEndReached={fetchNextRows}
-      />
-    ),
-    [dataLength, renderRow, fetchNextRows, minRowHeight, size]
-  );
+    return data;
+  };
 
   return (
-    <div
-      className={styles['Grid-body']}
-      // ref={ref}
-    >
-      {isLoadingMore && <ProgressBar className="position-absolute" state="indeterminate" size="small" />}
-
-      {/* <VirtualScroll
-        key="virtual-scroll"
-        className={styles['Grid-body']}
-        buffer={5}
-        length={20}
-        minItemHeight={minRowHeight[size]}
-        totalLength={dataLength}
-        renderItem={renderRow}
-        onEndReached={fetchNextRows}
-        // currentPage={currentPage}
-      /> */}
-      {memoizedVirtualScroll}
-
-      {isLoadingMore && <ProgressBar className="position-absolute bottom-0" state="indeterminate" size="small" />}
+    <div className={styles['Grid-body']}>
+      {getArrayList().map((item, i) => {
+        return renderRow(i, item);
+      })}
     </div>
   );
 };
